suggestions integrated
