+++ tests/util/test_fof_utils.py
@@ -9,15 +9,12 @@ from util.fof_utils import (
     clean_value,
     compare_arrays,
     compare_var_and_attr_ds,
-    expand_zip,
     get_observation_variables,
     get_report_variables,
     prepare_array,
     primary_check,
     print_entire_line,
     split_feedback_dataset,
-    to_list,
-    value_list,
     write_lines,
 )
 
@@ -353,16 +350,6 @@ def fixture_sample_dataset_3(sample_dataset_fof):
 #     assert (total, equal) == (108, 25)
 
 
-def test_value_list():
-    """
-    Test that the correct placeholder is given according to the different cases.
-    """
-    placeholders = {"fof": False}
-    assert value_list("fof", [1, 2, 3], placeholders) == [""]
-    placeholders = {"fof": True}
-    assert value_list("fof", [1, 2, 3], placeholders) == [1, 2, 3]
-
-
 # def test_get_file_type(tmp_path):
 #     """
 #     Test that the file is recognized as FOF or STAT if the corresponding keyword
@@ -396,74 +383,3 @@ def test_primary_check(tmp_path):
 
     false_result = primary_check(test_fof1, test_fof3)
     assert false_result is False
-
-
-@pytest.fixture(name="none_value", scope="function")
-def fixture_none_value():
-    return
-
-
-@pytest.fixture(name="string", scope="function")
-def fixture_string():
-    return "a,b,c,  ,e"
-
-
-@pytest.fixture(name="simple_list", scope="function")
-def fixture_list():
-    return [1, 2, 3, 4]
-
-
-def test_to_list_empty(none_value):
-    """
-    Test that the function returns an empty list if no value is given.
-    """
-    empty_return = to_list(none_value)
-    assert empty_return == []
-
-
-def test_to_list_full(string):
-    """
-    Test that if the function receives a string, it is split at each comma and
-    unnecessary spaces are removed.
-    """
-    string_list_to_list = to_list(string)
-    assert string_list_to_list == ["a", "b", "c", "e"]
-
-
-def test_to_list_no_changes(simple_list):
-    """
-    Tests that if the function receives a value that is neither a string
-    nor None, nothing changes.
-    """
-    to_list_no_changes = to_list(simple_list)
-    assert to_list_no_changes == [1, 2, 3, 4]
-
-
-# def test_expand_zip():
-#     """
-#     Test that the zip is expanded correctly.
-#     """
-#     zipped = [
-#         "test_{fof_type}.nc",
-#         "test_{fof_type}_{member_id}.nc",
-#         "test_{member_id}.nc",
-#     ]
-#     fof_type = ["AIREP", "PILOT"]
-#     member_ids = [1, 2]
-#     expanded_zip1 = expand_zip(zipped, fof_type, member_ids, member_type=None)
-#     expanded_zip2 = expand_zip(zip(zipped), fof_type, member_ids, member_type=None)
-#     assert expanded_zip1, expanded_zip2 == [
-#         "test_AIREP.nc",
-#         "test_PILOT.nc",
-#         "test_AIREP_1.nc",
-#         "test_AIREP_2.nc",
-#         "test_PILOT_1.nc",
-#         "test_PILOT_2.nc",
-#         "test_1.nc",
-#         "test_2.nc",
-#     ]
-
-#     expanded_zip3 = expand_zip(
-#         "test_{fof_type}.nc", fof_type, member_ids, member_type=None
-#     )
-#     assert expanded_zip3 == ["test_AIREP.nc", "test_PILOT.nc"]
