+++ util/fof_utils.py
@@ -4,8 +4,6 @@ This module contains functions for handling fof files
 
 import os
 import shutil
-from dataclasses import dataclass
-from enum import Enum
 
 import numpy as np
 import pandas as pd
@@ -358,127 +356,3 @@ def primary_check(file1, file2):
     name2_core = name2.replace("fof", "").replace(".nc", "")
 
     return name1_core == name2_core
-
-
-def to_list(value):
-    """
-    This function ensures that whatever value is passed to expand_zip,
-    one always gets something iterable and consistent.
-    """
-    if value is None:
-        return []
-    if isinstance(value, str):
-        return [v.strip() for v in value.split(",") if v.strip()]
-    return value
-
-
-def value_list(placeholder, values, placeholders):
-    """
-    Decide which values to use for a placeholder.
-    - If the placeholder is not present, returns [""]
-    - If placeholder is present but no values provided, returns [None]
-    """
-    if not placeholders.get(placeholder, False):
-        return [""]
-    return values or [None]
-
-
-def expand_zip(zipped, fof_type=None, member_ids=None, member_type=None):
-    """
-    Expands a list of tuples (zip) by replacing the placeholders
-    {fof_type}, {member_id}, and {member_type} with the values provided.
-    If a placeholder is present but no corresponding value is given,
-    the placeholder is left unchanged.
-    """
-
-    fof_list = to_list(fof_type)
-    member_list = to_list(member_ids)
-    member_type_list = to_list(member_type)
-
-
-    if isinstance(zipped, zip):
-        zipped = list(zipped)
-    elif isinstance(zipped, (list, tuple)):
-        zipped = [(z,) if not isinstance(z, tuple) else z for z in zipped]
-    else:
-        zipped = [(zipped,)]
-
-    expanded = []
-
-    for items in zipped:
-
-        file_info = FileInfo(items[0])
-        placeholders = {
-            key: any(f"{{{key}}}" in str(item) for item in items)
-            for key in ("fof_type", "member_id", "member_type")
-        }
-
-        fof_values = value_list("fof_type", fof_list, placeholders)
-        member_values = value_list("member_id", member_list, placeholders)
-        member_type_values = value_list("member_type", member_type_list, placeholders)
-
-        member_values_expanded = []
-
-
-        if file_info.type is FileType.STATS and member_type_list:
-            for m_id in member_values:
-                for m_type in member_type_list:
-                    member_values_expanded.append(f"{m_type}_{m_id}")
-        else:
-            member_values_expanded = member_values.copy()
-
-        for fof_val in fof_values:
-            for member_val in member_values_expanded:
-                formatted = [
-                        item.format(
-                            fof_type=fof_val or "{fof_type}",
-                            member_id=member_val or "{member_id}"
-                        )
-                        for item in items
-                    ]
-                expanded.append(
-                        formatted[0] if len(formatted) == 1 else tuple(formatted)
-                    )
-
-    return expanded
-
-
-class FileType(Enum):
-    """
-    Class that memorizes the distinction between file types
-    """
-
-    FOF = "fof"
-    STATS = "csv"
-
-
-@dataclass
-class FileInfo:
-    """
-    Class that memorize the path and the type of a file.
-    """
-
-    path: str
-    type: FileType = None
-
-    def __post_init__(self):
-
-        name = self.path.lower()
-
-        if "fof" in name:
-            self.type = FileType.FOF
-            return
-        if "csv" in name or "stats" in name:
-            self.type = FileType.STATS
-            return FileType.STATS
-
-        try:
-            with open(self.path, "r", encoding="utf-8") as f:
-                first_line = f.readline()
-                if "," in first_line or ";" in first_line:
-                    self.type = FileType.STATS
-                    return
-        except (OSError, FileNotFoundError):
-            pass
-
-        raise ValueError(f"Unknown file type for '{self.path}'")
